+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added: `selector-pseudo-class-blacklist` rule.
 - Added: `selector-pseudo-class-whitelist` rule.
 - Added: `media-feature-name-case` rule.
+- Added: regex support to the `ignore*` secondary options of the `*-no-unknown` rules.
 - Fixed: `media-feature-no-missing-punctuation` now ignores media features containing complex values e.g. `(min-width: ($var - 20px))` and `(min-width: calc(100% - 20px))`.
 - Fixed: `no-descending-specificity` message to correctly show which selector should come first.
 - Fixed: `selector-combinator-space-after` and `selector-combinator-space-before` now ignore operators within parenthetical non-standard constructs.